Configuration revamp pt 1
